Added my profile/work/about pages
Also added the divisors for the member information, but we need to add
the thumbnails, I'll start adding the css, i also might need you to
handle the emulation renders. I've just been using the chrome's
emulaters to handle the rendering. Not 100% sure if this would be
accurate. Hoping it is.